# specs/001-pos-inventory-ledger/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Moiz Mobile POS API
  version: 1.0.0
  description: >
    POS, inventory and customer ledger API for Moiz Mobile & Corporation, Danwran Lodhran.
    Every response uses the envelope described in conventions.md. All money is decimal with
    2 places; the server always recomputes totals from line items and ignores client totals.
servers:
  - url: /api
...
security:
  - bearerAuth: []
paths:
...
  /suppliers:
    get:
      tags: [Suppliers]
      summary: List suppliers with payable balances (Admin only)
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200': { description: Paged suppliers }
    post:
      tags: [Suppliers]
      summary: Create a supplier (Admin only)
      responses:
        '201': { description: Created }
  /suppliers/{id}:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    get:
      tags: [Suppliers]
      summary: Supplier detail with running payable
      responses:
        '200': { description: Supplier }
    put:
      tags: [Suppliers]
      summary: Update a supplier (Admin only)
      responses:
        '200': { description: Updated }
...
  /suppliers/{id}/payments:
    post:
      tags: [Suppliers]
      summary: Record a payment to a supplier (Admin only, FR-009)
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, paymentMethod]
              properties:
                amount: { type: number, format: double, minimum: 0.01 }
                paymentMethod: { $ref: '#/components/schemas/PaymentMethod' }
                note: { type: string, maxLength: 255 }
                confirmOverpayment:
                  type: boolean
                  default: false
                  description: Required true when amount exceeds the outstanding payable
      responses:
        '201': { description: Payment recorded }
        '400': { $ref: '#/components/responses/Error' }

  # ----------------------------------------------------------- Purchases
  /purchases:
    get:
      tags: [Purchases]
      summary: List purchases (Admin only)
      parameters:
        - { name: supplierId, in: query, schema: { type: integer } }
        - { $ref: '#/components/parameters/From' }
        - { $ref: '#/components/parameters/To' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200': { description: Paged purchases }
    post:
      tags: [Purchases]
      summary: Record a purchase (Admin only)
      description: >
        One transaction (FR-008, FR-011a): insert purchase, increase product stock,
        OVERWRITE products.costPrice with unitCost, increase supplier payable, write
        stock movement and audit rows. The new cost applies to all stock on hand.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [supplierId, productId, unitCost, quantity, purchaseDate]
              properties:
                supplierId: { type: integer }
                productId: { type: integer }
                unitCost: { type: number, format: double, minimum: 0.0001 }
                quantity: { type: integer, minimum: 1 }
                purchaseDate: { type: string, format: date-time }
                newSalePrice:
                  type: number
                  format: double
                  description: >
                    Optional. When supplied, updates products.salePrice, which then applies
                    to all remaining stock of that product (FR-011d).
      responses:
        '201': { description: Purchase recorded }
        '400': { $ref: '#/components/responses/Error' }

  # ----------------------------------------------------------- Customers
...
  /expenses:
    get:
      tags: [Expenses]
      summary: List expenses (Admin only)
      parameters:
        - { name: categoryId, in: query, schema: { type: integer } }
        - { $ref: '#/components/parameters/From' }
        - { $ref: '#/components/parameters/To' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200': { description: Paged expenses }
    post:
      tags: [Expenses]
      summary: Record an expense (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [categoryId, amount, expenseDate]
              properties:
                categoryId: { type: integer }
                amount: { type: number, format: double, minimum: 0.01 }
                expenseDate: { type: string, format: date-time }
                note: { type: string, maxLength: 255 }
      responses:
        '201': { description: Created }
...
  /dashboard:
    get:
      tags: [Reports]
      summary: Dashboard KPIs (Admin only, FR-035)
      parameters:
        - name: period
          in: query
          required: true
          schema: { type: string, enum: [Today, ThisMonth, ThisYear] }
      responses:
        '200':
          description: KPI set
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/DashboardDto' }
  /reports/sales:
    get:
      tags: [Reports]
      summary: Sales report grouped by day, month or year (Admin only)
      parameters:
        - { $ref: '#/components/parameters/From' }
        - { $ref: '#/components/parameters/To' }
        - name: groupBy
          in: query
          schema: { type: string, enum: [Day, Week, Month, Year], default: Day }
      responses:
        '200': { description: Sales rollup }
  /reports/profit:
    get:
      tags: [Reports]
      summary: Gross and net profit for a period (Admin only, FR-032)
      parameters:
        - { $ref: '#/components/parameters/From' }
        - { $ref: '#/components/parameters/To' }
        - name: groupBy
          in: query
          schema: { type: string, enum: [Day, Week, Month, Year], default: Month }
      responses:
        '200': { description: Profit rollup }
  /reports/profit-by-product:
    get:
      tags: [Reports]
      summary: Product-wise profit — total sale, total cost, total profit (Admin only, FR-033)
      parameters:
        - { $ref: '#/components/parameters/From' }
        - { $ref: '#/components/parameters/To' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200': { description: Paged product profit }
  /reports/stock:
    get:
      tags: [Reports]
      summary: Current stock report (Admin only)
      parameters:
        - { name: lowStockOnly, in: query, schema: { type: boolean, default: false } }
      responses:
        '200': { description: Stock report }
  /reports/stock-movements:
    get:
      tags: [Reports]
      summary: Full stock movement history across products (Admin only, FR-036)
      parameters:
        - { $ref: '#/components/parameters/From' }
        - { $ref: '#/components/parameters/To' }
        - { name: reason, in: query, schema: { type: string } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
      responses:
        '200': { description: Paged movements }
...
  /reports/expenses:
    get:
      tags: [Reports]
      summary: Expense report by category and period (Admin only)
      parameters:
        - { $ref: '#/components/parameters/From' }
        - { $ref: '#/components/parameters/To' }
      responses:
        '200': { description: Expense rollup }
  /reports/purchases:
    get:
      tags: [Reports]
      summary: Total purchases for a period (Admin only)
      parameters:
        - { $ref: '#/components/parameters/From' }
        - { $ref: '#/components/parameters/To' }
      responses:
        '200': { description: Purchase rollup }

  # ----------------------------------------------------------- Documents
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    From:
      name: from
      in: query
      schema: { type: string, format: date-time }
    To:
      name: to
      in: query
      schema: { type: string, format: date-time }
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
...
  schemas:
    Envelope:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        data: { nullable: true }
        error:
          nullable: true
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            traceId: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
    PaymentMethod:
      type: string
      enum: [Cash, BankTransfer, JazzCash, EasyPaisa, Raast, Credit, Partial]
    ProductStaffDto:
      type: object
      description: Staff-facing product. Has NO costPrice or profit property by design (FR-040).
      properties:
        id: { type: integer }
        name: { type: string }
        category: { type: string }
        brand: { type: string, nullable: true }
        model: { type: string, nullable: true }
        barcode: { type: string, nullable: true }
        imagePath: { type: string, nullable: true }
        salePrice: { type: number, format: double }
        quantityOnHand: { type: integer }
        isLowStock: { type: boolean }
...
    DashboardDto:
      type: object
      properties:
        period: { type: string, enum: [Today, ThisMonth, ThisYear] }
        totalSales: { type: number, format: double }
        totalPurchases: { type: number, format: double }
        grossProfit: { type: number, format: double }
        totalExpenses: { type: number, format: double }
        netProfit: { type: number, format: double }
        cashSales: { type: number, format: double }
        creditSales: { type: number, format: double }
        totalReceivables: { type: number, format: double }
        totalPayables: { type: number, format: double }
        itemsSoldCount: { type: integer }
        lowStockProducts:
          type: array
          items: { $ref: '#/components/schemas/ProductStaffDto' }